Removed setting the rmgrc file for benchmark and test builds
Previously the rmgrc file which directs RMG-Py to the appropriate location of RMG-database was stored in /HOME/.rmg/.rmgrc. This is dangerous because if a local run of RMG-tests does not complete, the rmgrc will still point to either the benchmark or test version of RMG-database, possibly without the user's knowledge. Afterwards, the user may run RMG and unintentionally use the wrong pathway to RMG-database.

It was found that RMG-Py will default to looking for RMG-database relative to its own position if nothing is provided. Because the local run of RMG-tests naturally sets up RMG-database this way for both bechmark and test builds, it is not necessary to ever set the rmgrc.